moduling the application as reviewed by the tse
